Nighthawk-PoC: refactor/clean
