+++ source/exe/client.cc
@@ -18,6 +18,7 @@
 #include "exe/benchmark_client.h"
 #include "exe/rate_limiter.h"
 #include "exe/sequencer.h"
+#include "exe/streaming_stats.h"
 
 using namespace std::chrono_literals;
 
@@ -48,72 +49,6 @@ uint32_t determine_cpu_cores_with_affinity() {
 }
 
 } // namespace
-
-StreamingStats::StreamingStats() { clear(); }
-
-void StreamingStats::clear() {
-  n = 0;
-  M1 = M2 = M3 = M4 = 0.0;
-}
-
-void StreamingStats::push(int64_t x) {
-  double delta, delta_n, delta_n2, term1;
-
-  int64_t n1 = n;
-  n++;
-  delta = x - M1;
-  delta_n = delta / n;
-  delta_n2 = delta_n * delta_n;
-  term1 = delta * delta_n * n1;
-  M1 += delta_n;
-  M4 += term1 * delta_n2 * (n * n - 3 * n + 3) + 6 * delta_n2 * M2 - 4 * delta_n * M3;
-  M3 += term1 * delta_n * (n - 2) - 3 * delta_n * M2;
-  M2 += term1;
-}
-
-int64_t StreamingStats::num_values() const { return n; }
-
-double StreamingStats::mean() const { return M1; }
-
-double StreamingStats::variance() const { return M2 / (n - 1.0); }
-
-double StreamingStats::stdev() const { return sqrt(variance()); }
-
-double StreamingStats::skewness() const { return sqrt(double(n)) * M3 / pow(M2, 1.5); }
-
-double StreamingStats::kurtosis() const { return double(n) * M4 / (M2 * M2) - 3.0; }
-
-StreamingStats operator+(const StreamingStats a, const StreamingStats b) {
-  StreamingStats combined;
-
-  combined.n = a.n + b.n;
-
-  double delta = b.M1 - a.M1;
-  double delta2 = delta * delta;
-  double delta3 = delta * delta2;
-  double delta4 = delta2 * delta2;
-
-  combined.M1 = (a.n * a.M1 + b.n * b.M1) / combined.n;
-
-  combined.M2 = a.M2 + b.M2 + delta2 * a.n * b.n / combined.n;
-
-  combined.M3 = a.M3 + b.M3 + delta3 * a.n * b.n * (a.n - b.n) / (combined.n * combined.n);
-  combined.M3 += 3.0 * delta * (a.n * b.M2 - b.n * a.M2) / combined.n;
-
-  combined.M4 = a.M4 + b.M4 +
-                delta4 * a.n * b.n * (a.n * a.n - a.n * b.n + b.n * b.n) /
-                    (combined.n * combined.n * combined.n);
-  combined.M4 += 6.0 * delta2 * (a.n * a.n * b.M2 + b.n * b.n * a.M2) / (combined.n * combined.n) +
-                 4.0 * delta * (a.n * b.M3 - b.n * a.M3) / combined.n;
-
-  return combined;
-}
-
-StreamingStats& StreamingStats::operator+=(const StreamingStats& rhs) {
-  StreamingStats combined = *this + rhs;
-  *this = combined;
-  return *this;
-}
 
 ClientMain::ClientMain(int argc, const char* const* argv) : ClientMain(OptionsImpl(argc, argv)) {}
 
@@ -227,7 +162,7 @@ bool ClientMain::run() {
                                       &sequencer](std::chrono::nanoseconds latency) {
         ASSERT(latency.count() > 0);
         results.push_back(latency.count());
-        streaming_stats.push(latency.count());
+        streaming_stats.addValue(latency.count());
 
         // Report results from the first worker about every one second.
         // TODO(oschaaf): we should only do this in explicit verbose mode because
@@ -236,13 +171,13 @@ bool ClientMain::run() {
         // influence timing of this happening.
         if (((results.size() % options_.requests_per_second()) == 0) && i == 0) {
           ENVOY_LOG(info,
-                    "#{} completions/sec. mean: {}+/-{}us. skewness: {}, kurtosis: {}."
+                    "#{} completions/sec. mean: {}+/-{}us. "
                     "pool connect failures: {}, overflow failures: {}. Replies: Good {}, Bad: "
                     "{}. Stream resets: {}.",
                     sequencer.completions_per_second(),
                     (static_cast<int64_t>(streaming_stats.mean())) / 1000,
                     (static_cast<int64_t>(streaming_stats.stdev())) / 1000,
-                    streaming_stats.skewness(), streaming_stats.kurtosis(),
+
                     client->pool_connect_failures(), client->pool_overflow_failures(),
                     client->http_good_response_count(), client->http_bad_response_count(),
                     client->stream_reset_count());
